cloud_function/requirements.txt
functions-framework==3.*
Flask==2.3.2
flask-cors==4.0.0
google-auth==2.23.4
requests==2.31.0
gunicorn==20.1.0
